python/firestore_read/google_cloud_run/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
google-cloud-firestore==2.5.2
